Update setup.py
psycopg2 keeps failing because it sees the latest version of postgresql as 10.5 and cannot interpret that as a version number correctly. 

Changed the version of psycopg2 to 2.7.5 as per https://github.com/psycopg/psycopg2/issues/489 because the maintainers have no desire to backport the fix. Here's hoping it doesn't fail now for some other reason on postgresql 10.5